+++ components/catalogs/back/tests/Integration/Infrastructure/Validation/ProductMapping/AttributeSource/SystemAttributeCategoriesSourceTest.php
@@ -46,6 +46,22 @@ class SystemAttributeCategoriesSourceTest extends AbstractAttributeSourceTest
         $this->assertEmpty($violations);
     }
 
+    public function testItReturnsNoViolationWithDefaultValue(): void
+    {
+        $source = [
+            'source' => 'categories',
+            'scope' => null,
+            'locale' => null,
+            'default' => 'Default value',
+            'parameters' => [
+                'label_locale' => 'en_US',
+            ],
+        ];
+        $violations = $this->validator->validate($source, new SystemAttributeCategoriesSource());
+
+        $this->assertEmpty($violations);
+    }
+
     /**
      * @dataProvider invalidDataProvider
      */
@@ -195,6 +211,18 @@ class SystemAttributeCategoriesSourceTest extends AbstractAttributeSourceTest
                     ],
                 ],
                 'expectedMessage' => 'This locale is disabled or does not exist anymore. Please check your channels and locales settings.',
+            ],
+            'invalid default value type' => [
+                'source' => [
+                    'source' => 'categories',
+                    'scope' => null,
+                    'locale' => null,
+                    'parameters' => [
+                        'label_locale' => 'en_EN',
+                    ],
+                    'default' => true,
+                ],
+                'expectedMessage' => 'This value should be of type string.',
             ],
         ];
     }