+++ front-packages/akeneo-design-system/src/components/SectionTitle/SectionTitle.tsx
@@ -94,6 +94,7 @@ const SectionTitle = ({children, ...rest}: SectionTitleProps) => {
         level: 'tertiary',
         size: 'small',
         ghost: true,
+        ...child.props
       });
     }