+++ src/Akeneo/Pim/Enrichment/Component/Product/Value/ScalarValue.php
@@ -55,14 +55,16 @@ class ScalarValue extends AbstractValue implements ValueInterface
             return false;
         }
 
+        $comparison = $value->getData() === $this->getData();
         if (\is_numeric($value->getData()) && \is_numeric($this->getData())) {
-            return $this->getScopeCode() === $value->getScopeCode() &&
-                $this->getLocaleCode() === $value->getLocaleCode() &&
-                bccomp($value->getData(), $this->getData()) === 0;
+            try {
+                $comparison = bccomp($value->getData(), $this->getData()) === 0;
+            } catch (\Throwable) {
+            }
         }
 
         return $this->getScopeCode() === $value->getScopeCode() &&
             $this->getLocaleCode() === $value->getLocaleCode() &&
-            $value->getData() === $this->getData();
+            $comparison;
     }
 }